refactor! computeCompletions now computes weighted at each level
This is a significant change in that it now requires tagWeights
at all levels on the metrics, tags on the entities, and missing
parameters on all the rules.

This still needs to be tested and a few rough edges smoothed,
particularly (in the latter case) surrounding missing data.

BREAKING CHANGE